fix test package name
