+++ seatunnel-engine/seatunnel-engine-server/src/main/test/org/apache/seatunnel/engine/server/execution/FixedCallTestTimeTask.java
@@ -15,10 +15,7 @@
  * limitations under the License.
  */
 
-package execution;
-
-import org.apache.seatunnel.engine.server.execution.ProgressState;
-import org.apache.seatunnel.engine.server.execution.Task;
+package org.apache.seatunnel.engine.server.execution;
 
 import lombok.NonNull;
